fix(transition_frontier/sync): failure when epoch ledger and root ledger are the same
We call `TransitionFrontierSyncLedgerSnarkedService::hashes_set` for the same key for the same ledger. This can happen E.g. if root snarked ledger is the same as staking or next epoch ledger, in which case we will sync same ledger twice. That causes assertion to fail in `mask.set_cached_hash_unchecked`.

remove once we have an optimization to not sync same ledgers/addrs multiple times.