fix(p2p/webrtc): initial peers are gone from peers list so we never manage to reconnect to them if we need to
We have a limit for max number of peers in peers list, once that is reached, peers are removed after disconnection. This is a problem because we were relying on those seed nodes staying in disconnected state so we could reconnect. To fix it, use list in `config.initial_peers` to reconnect to them even if we don't have them in peers list.